Split MICommandFactory.cs
MICommandFactory.cs was one large file which was getting difficult to edit. Ths splits it out to multiple files.
